Check if file exist before backup

+++ install.sh
@@ -5,6 +5,8 @@ FILES='zshrc zshenv gitconfig gitignore_global'
 
 for f in $FILES; do
     rm -f ~/.$f.orig
-    cp -f -R ~/.$f{,.orig}
+    if [ -f ~/.$f ]; then
+        cp -f -R ~/.$f{,.orig}
+    fi
     ln -v -s -f $CWD/$f ~/.$f
 done
